Move all deps to conda-forge, pin cuDNN <9.11, re-lock (v7), document GPU install
- jax/equinox/optax/numpyro now come from conda-forge like everything else;
  the only PyPI item left is the editable mrigen install (pip-only concept).
- The gpu environment uses conda-forge's CUDA build of jaxlib with
  cudnn <9.11, so pre-Turing GPUs (Pascal/Volta) keep working (fixes #5).
- pixi.lock regenerated with pixi 0.80 (lock format v7, all packages
  refreshed); CI pixi pinned to v0.80.0 to match.
- README gains a 'Running on a GPU' section for students.
- Fix 8 findings from the newer ruff this pulls in.

+++ src/mrigen/masks.py
@@ -31,7 +31,7 @@ def acs_columns(width: int, acs_frac: float) -> np.ndarray:
 
     Returns the integer column indices of the fully-sampled centre band.
     """
-    n_acs = max(1, int(round(acs_frac * width)))
+    n_acs = max(1, round(acs_frac * width))
     start = (width - n_acs) // 2
     return np.arange(start, start + n_acs)
